requirements.txt
Django==5.0.2
Pillow==9.5.0
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
python-dotenv==1.0.1
django-allauth==0.61.1
stripe==7.12.0
whitenoise==6.9.0 